requirements.txt
pymupdf==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3